Fix calculation of penalties
Var earnedTACoRewards is suposed to copy the values of potential rewards
and, later, modify the values. But up to now, this is not a deep-copy,
but a copy of the references of the objects contained there.

We are make sure we copy the values one by one and not just the
references.

Also, the function has been renamed.