Update subgraph's URL to the new version
The new subgraph version includes the "AuthorizationDecreaseRequested"
events as TACoAuthHistories.

The actual change here is that, from now on, when a staker requests to
decrease their TACo authorization, the script will treat the authorized
amount as the post-decrease value, i.e. the rewards will be calculated
from this value.

In oposition to the previous behaviour, in which the script considered
that the authorization decrease is effective when a
"AuthorizationDecreaseApproved" event is received.